minor changes and improvements in security

+++ src/main/java/pl/petersen/cinemamanager/service/HallService.java
@@ -3,25 +3,25 @@ package pl.petersen.cinemamanager.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.petersen.cinemamanager.entity.Hall;
+import pl.petersen.cinemamanager.entity.Seance;
 import pl.petersen.cinemamanager.entity.Seat;
 import pl.petersen.cinemamanager.repository.HallRepository;
 import pl.petersen.cinemamanager.repository.SeatRepository;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 
 @Service
 public class HallService {
 
     private final HallRepository hallRepository;
     private final SeatRepository seatRepository;
+    private final SeanceService seanceService;
 
     @Autowired
-    public HallService(HallRepository hallRepository, SeatRepository seatRepository) {
+    public HallService(HallRepository hallRepository, SeatRepository seatRepository, SeanceService seanceService) {
         this.hallRepository = hallRepository;
         this.seatRepository = seatRepository;
+        this.seanceService = seanceService;
     }
 
 
@@ -56,8 +56,8 @@ public class HallService {
 
 
     private Seat checkSeatExistence(Character letter, Integer num) {
-        if(seatRepository.count() > 0) {
-        List<Seat> existingSeats = getSeats();
+        if (seatRepository.count() > 0) {
+            List<Seat> existingSeats = getSeats();
             for (Seat existingSeat : existingSeats) {
                 if (existingSeat.getRow().equals(letter)
                         && existingSeat.getNumber().equals(num)) {
@@ -74,7 +74,12 @@ public class HallService {
         return hallRepository.findAll();
     }
 
-    public void deleteById(Long deleteId) {
-        hallRepository.deleteById(deleteId);
+    public Boolean deleteById(Long deleteId) {
+        long count = seanceService.countByHallId(deleteId);
+        if (count == 0) {
+            hallRepository.deleteById(deleteId);
+            return true;
+        }
+        return false;
     }
 }
