+++ src/main/java/pl/petersen/cinemamanager/service/MovieService.java
@@ -16,32 +16,17 @@ import java.util.Optional;
 public class MovieService {
 
     private final MovieRepository movieRepository;
+    private final SeanceService seanceService;
 
     @Autowired
-    public MovieService(MovieRepository movieRepository) {
+    public MovieService(MovieRepository movieRepository, SeanceService seanceService) {
         this.movieRepository = movieRepository;
+        this.seanceService = seanceService;
     }
 
-
-    public void save(Movie movie, MultipartFile poster) {
-        if (!poster.isEmpty()) {
-            setPoster(movie, poster);
-        }
+    public void save(Movie movie) {
         movieRepository.save(movie);
     }
-
-    private Movie setPoster(Movie movie, MultipartFile poster) {
-        Path newFile = Paths.get("src/main/webapp/uploads/posters/" +
-                movie.getTitle().replaceAll(" ", "_") + ".jpg");
-        try {
-            poster.transferTo(newFile);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        movie.setPoster(newFile.toString().substring(15));
-        return movie;
-    }
-
 
     public List<Movie> findAllOrderByNewest() {
         return movieRepository.findAllOrderCreatedOn();
@@ -51,8 +36,12 @@ public class MovieService {
         return movieRepository.findById(movieId);
     }
 
-    public void deleteById(Long id) {
-        movieRepository.deleteById(id);
+    public Boolean deleteById(Long id) {
+        if (seanceService.countByMovieId(id) == 0) {
+            movieRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
